Expose DefaultSchedulerService from backend-defaults
