+++ .changeset/great-clouds-hang.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-defaults': patch
+---
+
+Exposed `DefaultSchedulerService`
