+++ .changeset/shaggy-mugs-kiss.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-tasks': patch
+---
+
+More detailed deprecation messages
