+++ packages/backend-defaults/api-report-scheduler.md
@@ -3,8 +3,19 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { DatabaseService } from '@backstage/backend-plugin-api';
+import { LoggerService } from '@backstage/backend-plugin-api';
 import { SchedulerService } from '@backstage/backend-plugin-api';
 import { ServiceFactory } from '@backstage/backend-plugin-api';
+
+// @public
+export class DefaultSchedulerService {
+  // (undocumented)
+  static create(options: {
+    database: DatabaseService;
+    logger: LoggerService;
+  }): SchedulerService;
+}
 
 // @public
 export const schedulerServiceFactory: () => ServiceFactory<
