+++ packages/backend-defaults/src/entrypoints/scheduler/index.ts
@@ -14,4 +14,5 @@
  * limitations under the License.
  */
 
+export { DefaultSchedulerService } from './lib/DefaultSchedulerService';
 export { schedulerServiceFactory } from './schedulerServiceFactory';
